feat: add header helpers and extract into single file
