+++ src/utils/index.ts
@@ -9,3 +9,4 @@ export * from "./response";
 export * from "./session";
 export * from "./cors";
 export * from "./sanitize";
+export * from "./header";
